Updates to SJF and Process classes
This branch has updated the Algorithm abstract class to remove the "sortIOWaitQueue()" and has updated the comparator for the SJF class.

In doing so, I also updated the Process class, adding a method called "incrementBurstCycle()" to be called every time the CPU finishes with a burst cycle (not to be used when the I/O finishes a burst cycle).

Also in the Process class, I deleted some erroneous setter methods. I did not think setters were necessary for the "state" or "burstCycle" data members, as we have methods for setting the state to one of three specific states, and the burstCycle should only ever need to be incremented by one.

This branch is now the most complete.