Scheduler simulation working
A rough outline of the console program is working. Updates to the Scheduler class now get us incrementing the system timer, properly loading and dequeueing all queues, and properly executing both CPU and I/O bursts.

You can run the SimulatorClient program to see the movements of the processes from queue to CPU to I/O, etc, and to see the execution of the bursts. To slow down or speed up this action, just adjust the "Thread.sleep()" line of code at Line 94 of the SimulatorClient program.