Updates to the Scheduler Class
A couple updates. I had to reorder when I called the "executeIOBurst()" method within the "executeCPUBurst()" method in the Scheduler class to address a couple bugs that might arise in certain scenarios.

Also, I created a "printSchedulerQueues()" method in the SimulatorClient program to handle the output in the console so we can see what's going on as the simulator executes. It makes the "runScheduler()" method look a little less cluttered this way.