Updates to console output, and burst cycle logic
Ok, just a couple more updates to the console output (minor), and some updates to the logic in the Scheduler methods for handling the burst cycles.

Also, I updated the Scheduler to set the state of each process as it moves through the simulator, as well as the finishTime when the process terminates.