Updates to FCFS
FCFS algorithm working!

Deleted the "printResult()" abstract method from all Algorithm classes -- don't think it will be needed, but we can always add it back in later if necessary. Also, the "apply()" method does is it calls the "sortReadyQueue()" method. We may want to consider just nixing one of them.