NEW VERSION - DO NOT MERGE
DO NOT merge this branch to anything else. I ran into some problems getting Round Robin to work with the other program we had written, and so I created this branch to experiment with another method.

Instead of a separate class for Scheduler and an abstract Algorithm class, this program has an abstract Scheduler class which all algorithm classes will extend.

I have yet to get Round Robin working. However, so far FCFS, SJF, and Priority are working.

We will discuss when we get a chance.