Wait Time Updates
Updated the project so that wait times are incremented and displayed in the appropriate part of the program. Still need to check that they are accurate; will do that with old-fashioned pencil and paper.

Also, the "executeIO()" method is the same across all Scheduler classes, so I made it abstract. Also, the "executeCPU()" method is the same in all classes except RoundRobin, so I made that abstract as well and used @Override in the RoundRobin class to customize it for that algorithm.