Working!
OK, so updates were made so that all processes stay move through the Scheduler without ever leaving the JobQueue. This means there are various places where we needed to update the JobQueue with the current state of each process. But it works now, so instead of printing only the finished processes from the TerminatedProcesses list using the Logger, we can now just print the JobQueue. Note, I have not made those changes yet.