Manual Mode Update
Added loops in the SimulatorClient program in both the 'mainMenu()' and 'runScheduler()' methods to handle setting the simulator mode to either 'Manual' or 'Automatic'. Both seem to work in the couple tests I ran.

Also, when I tested RoundRobin, it wasn't working. Kept running on even when the last process had finished. Problem was that the changes that we made in the Scheduler class didn't apply to RoundRobin because RoundRobin was overriding those changes in the 'executeCPU()' method. I made the appropriate updates, and it works now.

Finally, I trimmed some fat in the Scheduler class by deleting the terminatedProcesses<> list since we don't technically need it now that we're just updating the jobQueue<> with the processes' states. I changed any references to the terminatedProcesses<> list to the jobQueue<> everywhere I found them in the program.